Continue with the Getting Started tutorial. Also some minor code changes.
Updates to getting started doc.

Minor variable name changes and variable typo fix.

Add the "Description" field to the output test report (after Dialog ID).

Turn on logging in CompareJObjects method when comparing expected to actual bot-replies. 

Logging is still off when CompareJObjects is used to check activities that we need to ignore.