More updates to README.md. Finished detailing all JSON fields
Finish detailing all the JSON fields in the app & test config file. I could not get a good compact view in a table format. Markdown is very limited on what you can do with tables. So I removed the table and discuss each JSON field in a separate section (title). This also means I can have a direct link to it when I need to reference a particular JSON field anywhere in the document.
What's left to do is work on the several "topic" sections at the bottom of the page.