Cancel ongoing audio output upon session start (button press, KWS)
Partners have requested information about how to implement "barge in;" more specifically, ensuring that a new interaction (e.g. via a keyword) can interrupt an ongoing existing interaction. This is a pretty simple thing to achieve and a decent default behavior, so it should be reflected in the Voice Assistant client.