Suppress error dialogs generated in response to graceful timeouts
There's a specific error that the service produces when a timeout gracefully disconnects via timeout (standard: 5 minutes). We currently present this error just like any other one, but that's causing a common misperception that something is wrong. This isn't an exceptional condition and so we shouldn't present it as such. The fix just suppresses the big red bubble and keeps a more useful explanation in the status box at the bottom of the tool.