Fix bug #255 by defining a state machine for KWS activation (#265)
Fix bug #255 - do not turn on KWS after TTS playback is done, if the user has previously selected to stop TTS.
This state machine will also help us easily implement optional barge-in when a configuration settings in the JSON indicates that barge-in needs to be supported (if the device has good echo cancellation). This is logged as feature request #262.
If the process I found out that if we are loading a wrong binary files as the KWS table, no error is generated. Logged as bug #264.