Fix & simplify UPL calculating in Voice Assistant Test
- There was a duplication of "Latency" values stored in BotReply objects and a separate UPL values which the BotConnector returns. Unify them and only calculate and report UPL.
- Update test configuration to use the expected UPL instead of just expected latency
- Simplify how the stopwatch is managed. No need to start/stop so many times. Restart once on SettionStrart, and then read elapsed time when an activity is received.

Hospitality WAV tests are failing, unrelated to the above changes, because for some reason it takes 4-5 seconds to get a reply in the test, even though when I do the same thing manually using Windows Voice Assistant Client everything runs fast. I'll file a bug on it.

Also, we need to wait and get elapsed time when the first TTS audio buffer is received, not when the activity is received. I'll open a work item on that.