fix(cpp-console): Removing the need to reinitialize ALSA after the first playback session. Initialising LinuxAudioPlayer m_shuttingDown flag to zero in constructor. removing the call to snd_pcm_close from the LinuxAudioPlayer's Stop method. (#519)
Hi there, I hope this isn't too presumptuous of me!
I've been playing around with the C++ example and I fixed a wee snag I hit with audio playback on Linux, I thought it was worth raising my fix as a PR to stop anyone else getting stuck.

I was trying out the cpp-console example client on a Raspberry Pi 3 Model B+ (on Ubuntu Server Ubuntu 20.04.1 LTS for ARM 64 bit) and I hit a few snags with audio playback.

The first issue was with m_shuttingDown in LinuxAudioPlayer not being initialised to zero, it was starting with a nonsense value. I realise the header file initialises it to zero, but the compiler that comes with Ubuntu doesn't respect it (I'm not sure what ought to happen, I'm just reporting what does happen).

The second issue was that PlayerThreadMain in LinuxAudioPlayer would always reinitialise ALSA for audio playback, this created a very long delay between DialogManager triggering playback and it actually happening. I've modified Stop so that it doesn't call snd_pcm_close (as I understand it, this should be called at the end of LinuxAudioPlayer's lifecycle) and replaced the call to Initialize with snd_pcm_prepare(m_playback_handle) which, according to my interpretation of the ALSA docs, should be all that's needed.